+++ mpp/include/mpp/PrimitiveModelStream.h
@@ -42,23 +42,23 @@ namespace mpp
 
 		std::map<std::string, size_t> getComponentOffsets(size_t& strideInBytes);
 
-		int getNumMeshes() const;
+		size_t getNumMeshes() const;
 
-		mesh::MeshSpecification const& getMeshSpecification(int meshIndex) const;
+		mesh::MeshSpecification const& getMeshSpecification(size_t meshIndex) const;
 
-		void getMeshCounts(int meshIndex, int* primitiveCount, int* vertexCount);
+		void getMeshCounts(size_t meshIndex, size_t* primitiveCount, size_t* vertexCount);
 
-		VertexDataStreamDefinition getMeshDataStream(int meshIndex, mesh::Vertex::Component component) const;
+		VertexDataStreamDefinition getMeshDataStream(size_t meshIndex, mesh::Vertex::Component component) const;
 
-		int getMeshIndexWidth(int meshIndex) const;
+		size_t getMeshIndexWidth(size_t meshIndex) const;
 
-		float getMeshPointSize(int meshIndex) const;
+		float getMeshPointSize(size_t meshIndex) const;
 
-		uint8_t const* getMeshIndexData(int meshIndex) const;
+		uint8_t const* getMeshIndexData(size_t meshIndex) const;
 
-		std::string const& getMeshName(int meshIndex) const;
+		std::string const& getMeshName(size_t meshIndex) const;
 
-		std::string const& getMeshMaterial(int meshIndex) const;
+		std::string const& getMeshMaterial(size_t meshIndex) const;
 
 		template<typename T>
 		void setVertexData(int offset, std::initializer_list<T> const& vertex)
